Split command registerring into a new file.
