+++ package.json
@@ -26,6 +26,7 @@
     "axios": "^1.5.1",
     "discord.js": "^14.13.0",
     "dotenv": "^16.3.1",
+    "fs": "^0.0.1-security",
     "nodemon": "^3.0.1"
   }
 }
